requirements.txt
pandas==1.0.5
numpy==1.18.5
streamlit==0.82.0
Pillow==8.2.0
